requirements.txt
django
requests
psycopg2
pylint
djangosslserver
django-pwa
python-decouple